faq-chatbot-rag/requirements.txt
streamlit>=1.38
sentence-transformers>=3.0
faiss-cpu>=1.8
pypdf>=4.3
openai>=1.40
torch>=2.3
numpy>=1.26